Analysis of Problem
Problem : 
A major travel company is experiencing troubles with providing customers with suitable relationship managers, that can help provide suitable holiday packages toward the customer.   

Objective: 
Our objective is to help improve the operation of the travel companies in house call management center. To achieve this, we will be designing and implementing a system that can help adjust the call flow rate, and match customers to suitable relationship managers that can provide the correct holiday packages to the customer.  

Stakeholder Identification:
• Relationship Managers
• Customers
• Business Executives
• System Developer